[Implementation] Implemented PopulateOptions, tweaked Board/Cell methods
Implemented the PopulateOptions helper algorithm in AbstractSolver, an
abstract class implementation of the Solver interface. This way any
solver algorithm can access the PopulateOptions helper method. Tweaked
some of the Cell and Board accessor methods to match outputs to
requisite input types for PopulateOptions. Calculated the time
complexity for the Prune-and-Conquer approach, O(n^4), which is bad but
still polytime which is the goal.

TODO: Implement Prune-and-Conquer algorithm, create input files, and then test to see if it actually works (spoiler: it probably doesn't but have to start somewhere)